+++ image_analogy/losses/patches.py
@@ -25,7 +25,7 @@ def reconstruct_from_patches_2d(patches, image_size):
     '''
     i_h, i_w = image_size[:2]
     p_h, p_w = patches.shape[1:3]
-    img = np.zeros(image_size)
+    img = np.zeros(image_size, dtype=np.float32)
     # compute the dimensions of the patches array
     n_h = i_h - p_h + 1
     n_w = i_w - p_w + 1
@@ -45,7 +45,7 @@ def combine_patches(patches, out_shape):
     '''Reconstruct an image from these `patches`'''
     patches = patches.transpose(0, 2, 3, 1)
     recon = reconstruct_from_patches_2d(patches, out_shape)
-    return recon.transpose(2, 0, 1)
+    return recon.transpose(2, 0, 1).astype(np.float32)
 
 
 def find_patch_matches(a, a_norm, b):
